wip: move out user helpers into their own file
